requirements.txt
crewai
langchain-openai
python-dotenv
gTTS
moviepy
pillow

# YouTube upload
google-api-python-client==2.93.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-auth==2.22.0
